[refactor] 홈 컴포넌트 분리 (#118)
