+++ src/pages/home/Home.css.ts
@@ -73,6 +73,7 @@ export const listWrapper = style({
 
 export const imgBanner = style({
   width: '100%',
+  aspectRatio: '4.27 / 1',
   marginBottom: '6.6rem',
 
   cursor: 'pointer',
